requirements-runtime.txt
# Core runtime
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.2.0
pyarrow==15.0.0

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database (optional in demo mode)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Inference + utilities
scikit-learn==1.4.0
joblib==1.3.2
pyyaml==6.0.1
click==8.1.7
httpx==0.26.0
requests==2.31.0
